Reorganising structure
Reorganised structure PLPVL (PresetLayoutPartVitaminLibary). We have
Presets / Settings / Renderer Layouts / Parts / Vitamins / Libraries.
Some of the parts are moved to the new structure like  Z-axes, and Y
axes parts and spacer, motor_mount libs. The rest of the files not yet
working in the new structure yet. Z-axes parts quiet mature, Y axes
parts is ok now. Spacer lib is done, but there is a bug in the
motor_mount with Nema_23 motors.